Use github's API html_url as the node's link href
Instead of manually recreating the href, we use the one given to us by
the API. This way, we stop assuming that each file lives inside
`github.com/...<repo>.../<filepath>`.

By doing this, we allow a file/dir inside the repo to being linked to
itself, and a submodule to being linked to the submodule's own
repository.